Main Postprocessing Script
This code performs the semi-autonomous full procedure for post-processing CT images collected from the modified CT scanner in the Petroleum and Geosystems Engineering Department at the University of Texas at Austin. The code begins by extracting the attenuation data from the raw data file output by the CT scanner. The data is then masked, filtered, and corrected to produce a reliable dataset for further analysis. The porosity, bulk density, and formation front parameters are calculated from the finalized dataset. The code has the ability to produce plots to visualize the data produced during post-processing, but that option can be turned off if desired. At no point during the process is any data overwritten. All iterations of the data from raw to corrected are saved in separate folders to ensure continuity and for error analysis.

All other scripts should be contained within a subfolder named "SUBROUTINES" in the same directory as the main script. The inputs required for all the subroutines should be created automatically within the main script. The data to be processed and any required input files are all described in the comments. Thus, the user to should read through the script carefully prior to usage to be sure all the required elements are in place.